requirements.txt
requests==2.26.0
flask==2.0.1
python_dotenv==0.19.0
pandas==1.3.2
sqlalchemy==1.4.23
psycopg2==2.9.1
gunicorn==20.1.0
flask-bootstrap==3.3.7.1